fix(grid): devolver items no absorbidos al jugador y prevenir overflow de energia

+++ src/main/java/io/github/sefiraat/networks/network/NetworkRoot.java
@@ -890,7 +890,7 @@ public class NetworkRoot extends NetworkNode {
             return;
         }
 
-        int removed = 0;
+        long removed = 0;
         for (Location node : powerNodes) {
             final SlimefunItem item = BlockStorage.check(node);
             if (item instanceof NetworkPowerNode powerNode) {
